Refactor: move /certmanager into /pkg
Specifically, CA related methods are moved into /pkg/pki/ca; GCP token
fetcher is moved into /pkg/credential/
