+++ cmd/istio_ca/BUILD
@@ -5,9 +5,9 @@ go_library(
     srcs = ["main.go"],
     visibility = ["//visibility:private"],
     deps = [
-        "//certmanager:go_default_library",
         "//cmd/istio_ca/version:go_default_library",
         "//controller:go_default_library",
+        "//pkg/pki/ca:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
