+++ controller/BUILD
@@ -5,8 +5,8 @@ go_library(
     srcs = ["secret.go"],
     visibility = ["//visibility:public"],
     deps = [
-        "//certmanager:go_default_library",
         "//pkg/pki:go_default_library",
+        "//pkg/pki/ca:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
@@ -25,7 +25,7 @@ go_test(
     srcs = ["secret_test.go"],
     library = ":go_default_library",
     deps = [
-        "//certmanager:go_default_library",
+        "//pkg/pki/ca:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
         "@io_k8s_client_go//kubernetes/fake:go_default_library",
