+++ pkg/credential/token.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package certmanager
+package credential
 
 import (
 	"cloud.google.com/go/compute/metadata"
